refactor: Buyer, Seller 를 domain/subject/ 폴더로 이동
