+++ src/LottoController.js
@@ -1,6 +1,6 @@
-import Buyer from './domain/Buyer';
 import LottoResult from './domain/LottoResult';
-import Seller from './domain/Seller';
+import Buyer from './domain/subject/Buyer';
+import Seller from './domain/subject/Seller';
 import WinningLotto from './domain/WinningLotto';
 import InputView from './view/InputView';
 import OutputView from './view/OutputView';
